requirements.txt
Django==1.11.24
djangorestframework==3.6.2
pbr==5.4.2
pytz==2019.2
six==1.12.0
stevedore==1.30.1
virtualenv==16.7.4
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
uWSGI==2.0.15
psycopg2
dj-database-url==0.5.0
whitenoise==3.2.1
gunicorn==19.7.1
django-cors-headers